Remove symbolic link in example/.clang-format
Symbolic links are disabled by default in the Git for Windows
installer, so on Windows the file does not work as intended.

Furthermore, even if the symbolic link are enabled, the git symbolic
links on Windows are not correctly processed, and create spurious
changes in the Git repo even if it is not modified
(see https://github.com/robotology/robotology-superbuild/issues/429).

To avoid these class of problems, we just copy the root .clang-format
in example/.clang-format .